chat-with-your-data-using-chatgpt/requirements.txt
openai
python-dotenv
langchain
langchain-openai
pypdf
faiss-cpu
langchainhub
langchain-community